Title screen changes, some UI stuff,
Title screen is now updated to reflect the current assets. I also changed around the buttons and added a logo to the title screen. I added a button to access the credits, but it doesn't go anywhere. Home strech :D. I've still got to add the main menu music, but we are close to being done